get_next_line

+++ get_next_line.c
@@ -6,7 +6,7 @@
 /*   By: rkhakimu <rkhakimu@student.hive.fi>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/05/15 15:21:58 by rkhakimu          #+#    #+#             */
-/*   Updated: 2024/05/27 18:18:19 by rkhakimu         ###   ########.fr       */
+/*   Updated: 2024/06/11 18:50:22 by rkhakimu         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -85,7 +85,7 @@ char	*get_next_line(int fd)
 	static char	buffer[BUFFER_SIZE + 1];
 	char		*line;
 	char		*next_line;
-	int			bytes;
+	ssize_t		bytes;
 
 	line = NULL;
 	if (!check_input_gnl(fd, buffer))
